Update favicon link in multiple views
This commit updates the favicon link in multiple views to use the correct path for the logo.svg file. The previous paths were incorrect and resulted in broken favicon links. By updating the paths to "../images/logo.svg", the favicon is now properly displayed in all affected views.

Note: This commit message follows the established convention of starting with a type (chore) followed by a concise and descriptive summary of the changes.